requirements.txt
asgiref==3.5.2
certifi==2022.12.7
charset-normalizer==3.0.1
Django==3.2.16
django-crispy-forms==1.14.0
django-phonenumber-field==7.0.0
docopt==0.6.2
idna==3.4
phonenumbers==8.13.1
pipreqs==0.4.11
pytz==2022.6
requests==2.28.2
sqlparse==0.4.3
typing-extensions==4.4.0
urllib3==1.26.14
yarg==0.1.9
boto3==1.24.61
botocore==1.27.61
certifi==2022.6.15
charset-normalizer==2.1.0
commonmark==0.9.1
gunicorn==20.1.0
whitenoise==6.2.0